requirements.txt
numpy
pillow
requests
tqdm
matplotlib
opencv-python